requirements.txt
Django>=4.2
gunicorn>=23.0.0
python-dotenv~=1.1.0
Pillow
whitenoise
psycopg[binary,pool]
dj_database_url
requests
